refractor Command to own file
