+++ src/main/java/jiachen/duke/Command.java
@@ -1,0 +1,39 @@
+package jiachen.duke;
+
+/**
+ * The enum Command.
+ */
+enum Command {
+  /**
+   * Exit command command.
+   */
+  EXIT_COMMAND,
+  /**
+   * List command command.
+   */
+  LIST_COMMAND,
+  /**
+   * Done command command.
+   */
+  DONE_COMMAND,
+  /**
+   * Delete command command.
+   */
+  DELETE_COMMAND,
+  /**
+   * Todo command command.
+   */
+  TODO_COMMAND,
+  /**
+   * Deadline command command.
+   */
+  DEADLINE_COMMAND,
+  /**
+   * Event command command.
+   */
+  EVENT_COMMAND,
+  /**
+   * Find command command.
+   */
+  FIND_COMMAND
+}
